+++ Sources/RichTextKit/Sharing/StandardRichTextShareService.swift
@@ -1,6 +1,6 @@
 //
 //  StandardRichTextShareService.swift
-//  OribiWriter
+//  RichTextKit
 //
 //  Created by Daniel Saidi on 2021-12-06.
 //  Copyright © 2021 Daniel Saidi. All rights reserved.