+++ Sources/RichTextKit/Views/Buttons/RichTextStyleButton.swift
@@ -49,7 +49,7 @@ public struct RichTextStyleButton: View {
             style.icon
                 .foregroundColor(tintColor)
                 .contentShape(Rectangle())
-        }
+        }.keyboardShortcut(for: style)
     }
 }