+++ Tests/RichTextKitTests/RichTextCoordinator+SubscriptionsTests.swift
@@ -1,6 +1,6 @@
 //
 //  RichTextCoordinator+SubscriptionsTests.swift
-//  OribiRichTextKitTests
+//  RichTextKitTests
 //
 //  Created by Daniel Saidi on 2022-12-05.
 //  Copyright © 2022 Daniel Saidi. All rights reserved.